requirements.txt
click==6.7
Django==1.11.7
djangorestframework==3.7.3
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.12
pytz==2017.3
Werkzeug==0.12.2
